refactor: Consolidate theme management by removing themeStore and updating imports
