--- src/lib/themeStore.svelte.ts
@@ -1,4 +1,0 @@
-import { writable } from 'svelte/store';
-import { DEFAULT_THEME } from './themesTypes';
-
-export const currentTheme = writable(DEFAULT_THEME);
